Move function definitions to .cpp
